refactor(ecstore): isolate local object rename commit
